// frontend/graphql/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ISO8601DateTime: any;
};

export type Mutation = {
  __typename?: 'Mutation';
  createTag?: Maybe<CreateTagPayload>;
  deleteTag?: Maybe<DeleteTagPayload>;
  updateTag?: Maybe<UpdateTagPayload>;
};


export type MutationCreateTagArgs = {
  input: CreateTagInput;
};


export type MutationDeleteTagArgs = {
  input: DeleteTagInput;
};


export type MutationUpdateTagArgs = {
  input: UpdateTagInput;
};

export type Query = {
  __typename?: 'Query';
  tag: Tag;
  tags: Array<Tag>;
};


export type QueryTagArgs = {
  id: Scalars['ID'];
};

export type Tag = {
  __typename?: 'Tag';
  activeFlag: Scalars['Boolean'];
  createdAt: Scalars['ISO8601DateTime'];
  id: Scalars['ID'];
  name: Scalars['String'];
  tagNumber: Scalars['Int'];
  updatedAt: Scalars['ISO8601DateTime'];
};

export type TagInput = {
  activeFlag?: InputMaybe<Scalars['Boolean']>;
  id?: InputMaybe<Scalars['ID']>;
  name?: InputMaybe<Scalars['String']>;
  tagNumber?: InputMaybe<Scalars['Int']>;
};

/** Autogenerated input type of createTag */
export type CreateTagInput = {
  /** A unique identifier for the client performing the mutation. */
  clientMutationId?: InputMaybe<Scalars['String']>;
  params: TagInput;
};

/** Autogenerated return type of createTag. */
export type CreateTagPayload = {
  __typename?: 'createTagPayload';
  /** A unique identifier for the client performing the mutation. */
  clientMutationId?: Maybe<Scalars['String']>;
  tag: Tag;
};

/** Autogenerated input type of deleteTag */
export type DeleteTagInput = {
  /** A unique identifier for the client performing the mutation. */
  clientMutationId?: InputMaybe<Scalars['String']>;
  id: Scalars['ID'];
};

/** Autogenerated return type of deleteTag. */
export type DeleteTagPayload = {
  __typename?: 'deleteTagPayload';
  /** A unique identifier for the client performing the mutation. */
  clientMutationId?: Maybe<Scalars['String']>;
  deleted: Scalars['Boolean'];
};

/** Autogenerated input type of updateTag */
export type UpdateTagInput = {
  /** A unique identifier for the client performing the mutation. */
  clientMutationId?: InputMaybe<Scalars['String']>;
  params: TagInput;
};

/** Autogenerated return type of updateTag. */
export type UpdateTagPayload = {
  __typename?: 'updateTagPayload';
  /** A unique identifier for the client performing the mutation. */
  clientMutationId?: Maybe<Scalars['String']>;
  tag: Tag;
};